// libs/store/Auth/actions-type.tsx
export enum ActionType {
    AUTH_LOADING = 'auth/AUTH_LOADING',
    AUTH_ERROR = 'auth/AUTH_ERROR',
    CLEAR_AUTH_ERROR = 'auth/CLEAR_AUTH_ERROR',
    AUTH_SUCCESS = 'auth/AUTH_SUCCESS',
    REGISTER_COMPLETED = 'auth/REGISTER_COMPLETED',
    ACTIVATE_ACCOUNT = 'auth/ACTIVATE_ACCOUNT',
    SIGN_IN_WITH_EMAIL_AND_PASSWORD = 'auth/SIGN_IN_WITH_EMAIL_AND_PASSWORD',
    REGISTER_WITH_EMAIL_AND_PASSWORD = 'auth/REGISTER_WITH_EMAIL_AND_PASSWORD',
    REGISTER_OR_LOGIN_AS_FACEBOOK_USER = 'auth/REGISTER_OR_LOGIN_AS_FACEBOOK_USER',
    REGISTER_OR_LOGIN_AS_GOOGLE_USER = 'auth/REGISTER_OR_LOGIN_AS_GOOGLE_USER',
    REGISTER_OR_LOGIN_AS_TWITTER_USER = 'auth/REGISTER_OR_LOGIN_AS_TWITTER_USER',
    LOGOUT = 'auth/LOGOUT',
    INITIAL_LOAD = 'auth/INITIAL_LOAD',
    SEND_RESET_PASSWORD_EMAIL = 'auth/SEND_RESET_PASSWORD_EMAIL',
    RESET_PASSWORD = 'auth/RESET_PASSWORD',
    GET_SHOPS = 'auth/GET_SHOPS',
    GET_SHOPS_SUCCESS = 'auth/GET_SHOPS_SUCCESS',
    GET_SHOPS_FAILED = 'auth/GET_SHOPS_FAILED',
}